% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataconversn.R
\name{csv2JSON}
\alias{csv2JSON}
\title{csv file to JSON format}
\usage{
csv2JSON(input_file)
}
\arguments{
\item{input_file}{the path of a csv file to read the file from; Only one  file must be supplied}
}
\description{
Conversion of csv data to JSON format.
}
\examples{
test1 <- as.character(sample(1:100,10))
input_file <- tempfile(fileext = ".csv")
writeLines(test1, input_file)
csv2JSON(input_file = input_file)

}
