% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataconversn.R
\name{excel2JSON}
\alias{excel2JSON}
\title{Excel file to JSON format}
\usage{
excel2JSON(input_file)
}
\arguments{
\item{input_file}{the path of a excel file to read the file from; Only one  file must be supplied}
}
\description{
Conversion of excel data to JSON format.
}
