% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataconversn.R
\name{JSON2csv}
\alias{JSON2csv}
\title{JSON file to csv format}
\usage{
JSON2csv(input_file)
}
\arguments{
\item{input_file}{the path of a JSON file to read the file from; Only one  file must be supplied}
}
\description{
Conversion of JSON data to csv format.
}
\examples{
test1 <- '{"a":true, "b":false, "c":null}'
input_file <- tempfile(fileext = ".json")
writeLines(test1, input_file)
JSON2csv(input_file = input_file)

}
